Allow participants to register via invitations
This commit creates a new path for participants to register, via a link
provided to them by their coach or team-mate. When an institution
registers, they are given unique random codes that are used in query
parameters to authorize the use of the team/adjudicator forms (and then
as a hidden form input), which will be active until the registration
period is over.

On creating a team with incomplete speakers, a similar code will be
created for team-mates to individually submit their data. Notices are
also added to prompt the URL to be shared.